Move derive test into its own file
